refactor: 논리 흐름대로 main.js의 로직 순서 수정 (main.js)
가독성을 위하여 논리성 및 예측 가능성을 높이고자 로직 순서 변경
